+++ src/writer.c
@@ -61,10 +61,6 @@ struct signal_info_s {
     char units[128];
     struct track_info_s tracks[JLS_TRACK_TYPE_COUNT];   // array index is jls_track_type_e
     struct jls_wf_f32_s * signal_writer;
-
-    uint8_t data_tag;
-    uint8_t summary_tag;
-    uint8_t index_tag;
 };
 
 struct buf_s {
@@ -111,6 +107,17 @@ struct jls_signal_def_s SIGNAL_0 = {       // 0 reserved for VSR annotations
         .name = "global_annotation_signal",
         .units = "",
 };
+
+/**
+ * @brief construct the chunk tag.
+ *
+ * @param track_type The jls_track_type_e
+ * @param track_chunk The jls_track_chunk_e
+ * @return
+ */
+static inline uint8_t tag_construct(uint8_t track_type, uint8_t track_chunk) {
+    return 0x20 | ((track_type & 0x03) << 3) | (track_chunk & 0x07);
+}
 
 int32_t jls_wr_open(struct jls_wr_s ** instance, const char * path) {
     if (!instance) {
@@ -357,7 +364,7 @@ static int32_t track_wr_def(struct jls_wr_s * self, struct track_info_s * track_
     }
     chunk->hdr.item_next = 0;  // update later
     chunk->hdr.item_prev = self->signal_mra.offset;
-    chunk->hdr.tag = 0x20 | ((track_info->track_type & 0x03) << 3) | JLS_TRACK_CHUNK_DEF;
+    chunk->hdr.tag = tag_construct(track_info->track_type, JLS_TRACK_CHUNK_DEF);
     chunk->hdr.rsv0_u8 = 0;
     chunk->hdr.chunk_meta = track_info->signal_id;
     chunk->hdr.payload_length = 0;
@@ -376,7 +383,7 @@ static int32_t track_wr_head(struct jls_wr_s * self, struct track_info_s * track
     if (!chunk->offset) {
         chunk->hdr.item_next = 0;  // update later
         chunk->hdr.item_prev = self->signal_mra.offset;
-        chunk->hdr.tag = 0x20 | ((track_info->track_type & 0x03) << 3) | JLS_TRACK_CHUNK_HEAD;
+        chunk->hdr.tag = tag_construct(track_info->track_type, JLS_TRACK_CHUNK_HEAD);
         chunk->hdr.rsv0_u8 = 0;
         chunk->hdr.chunk_meta = track_info->signal_id;
         chunk->hdr.payload_length = sizeof(track_info->head_offsets);
@@ -450,18 +457,12 @@ int32_t jls_wr_signal_def(struct jls_wr_s * self, const struct jls_signal_def_s 
                 JLS_LOGE("FSR requires sample rate");
                 return JLS_ERROR_PARAMETER_INVALID;
             }
-            info->data_tag = JLS_TAG_TRACK_FSR_DATA;
-            info->summary_tag = JLS_TAG_TRACK_FSR_SUMMARY;
-            info->index_tag = JLS_TAG_TRACK_FSR_INDEX;
             break;
         case JLS_SIGNAL_TYPE_VSR:
             if (def->sample_rate) {
                 JLS_LOGE("VSR but sample rate specified, ignoring");
                 def->sample_rate = 0;
             }
-            info->data_tag = JLS_TAG_TRACK_VSR_DATA;
-            info->summary_tag = JLS_TAG_TRACK_VSR_SUMMARY;
-            info->index_tag = JLS_TAG_TRACK_VSR_INDEX;
             break;
         default:
             JLS_LOGE("Invalid signal type: %d", (int) def->signal_type);
@@ -509,6 +510,9 @@ int32_t jls_wr_signal_def(struct jls_wr_s * self, const struct jls_signal_def_s 
     ROE(update_mra(self, &self->signal_mra, chunk));
 
     if (def->signal_type == JLS_SIGNAL_TYPE_FSR) {
+        info->tracks[JLS_TRACK_TYPE_FSR].track_type = JLS_TRACK_TYPE_FSR;
+        info->tracks[JLS_TRACK_TYPE_ANNOTATION].track_type = JLS_TRACK_TYPE_ANNOTATION;
+        info->tracks[JLS_TRACK_TYPE_UTC].track_type = JLS_TRACK_TYPE_UTC;
         ROE(track_wr_def(self, &info->tracks[JLS_TRACK_TYPE_FSR]));
         ROE(track_wr_head(self, &info->tracks[JLS_TRACK_TYPE_FSR]));
         ROE(track_wr_def(self, &info->tracks[JLS_TRACK_TYPE_ANNOTATION]));
@@ -516,6 +520,8 @@ int32_t jls_wr_signal_def(struct jls_wr_s * self, const struct jls_signal_def_s 
         ROE(track_wr_def(self, &info->tracks[JLS_TRACK_TYPE_UTC]));
         ROE(track_wr_head(self, &info->tracks[JLS_TRACK_TYPE_UTC]));
     } else if (def->signal_type == JLS_SIGNAL_TYPE_VSR) {
+        info->tracks[JLS_TRACK_TYPE_VSR].track_type = JLS_TRACK_TYPE_VSR;
+        info->tracks[JLS_TRACK_TYPE_ANNOTATION].track_type = JLS_TRACK_TYPE_ANNOTATION;
         ROE(track_wr_def(self, &info->tracks[JLS_TRACK_TYPE_VSR]));
         ROE(track_wr_head(self, &info->tracks[JLS_TRACK_TYPE_VSR]));
         ROE(track_wr_def(self, &info->tracks[JLS_TRACK_TYPE_ANNOTATION]));
@@ -605,7 +611,7 @@ int32_t jls_wr_data_prv(struct jls_wr_s * self, uint16_t signal_id, enum jls_tra
 
     chunk.hdr.item_next = 0;  // update later
     chunk.hdr.item_prev = track->data.offset;
-    chunk.hdr.tag = info->data_tag;
+    chunk.hdr.tag = tag_construct(track->track_type, JLS_TRACK_CHUNK_DATA);
     chunk.hdr.rsv0_u8 = 0;
     chunk.hdr.chunk_meta = signal_id | (0 << 12);
     chunk.hdr.payload_length = payload_length;
@@ -640,7 +646,7 @@ int32_t jls_wr_summary_prv(struct jls_wr_s * self, uint16_t signal_id, enum jls_
 
     chunk.hdr.item_next = 0;  // update later
     chunk.hdr.item_prev = track->summary[level].offset;
-    chunk.hdr.tag = info->summary_tag;
+    chunk.hdr.tag = tag_construct(track->track_type, JLS_TRACK_CHUNK_SUMMARY);
     chunk.hdr.rsv0_u8 = 0;
     chunk.hdr.chunk_meta = signal_id | (((uint16_t) level) << 12);
     chunk.hdr.payload_length = payload_length;
@@ -671,7 +677,7 @@ int32_t jls_wr_index_prv(struct jls_wr_s * self, uint16_t signal_id, enum jls_tr
 
     chunk.hdr.item_next = 0;  // update later
     chunk.hdr.item_prev = track->index[level].offset;
-    chunk.hdr.tag = info->index_tag;
+    chunk.hdr.tag = tag_construct(track->track_type, JLS_TRACK_CHUNK_INDEX);
     chunk.hdr.rsv0_u8 = 0;
     chunk.hdr.chunk_meta = signal_id | (((uint16_t) level) << 12);;
     chunk.hdr.payload_length = payload_length;